@@ -9,8 +9,8 @@ apply plugin: 'idea'
 
 group = "ch.softappeal.yass"
 
-sourceCompatibility = '1.7'
-targetCompatibility = '1.7'
+sourceCompatibility = '1.8'
+targetCompatibility = '1.8'
 
 configurations.all {
     resolutionStrategy {
@@ -27,8 +27,8 @@ sourceSets {
 }
 
 def websocket_dep = 'javax.websocket:javax.websocket-api:1.0'
-def jetty_dep = 'org.eclipse.jetty.websocket:javax-websocket-server-impl:9.2.14.v20151106'
-def undertow_dep = 'io.undertow:undertow-websockets-jsr:1.3.5.Final'
+def jetty_dep = 'org.eclipse.jetty.websocket:javax-websocket-server-impl:9.3.6.v20151106'
+def undertow_dep = 'io.undertow:undertow-websockets-jsr:1.3.10.Final'
 def junit_dep = 'junit:junit:4.12'
 
 dependencies {
@@ -39,11 +39,12 @@ dependencies {
 
 tasks.withType(JavaCompile) {
     options.compilerArgs << "-Xlint"
+    options.compilerArgs << "-parameters"
     options.encoding = "US-ASCII"
 }
 
 task generateTypeScriptContract(type: JavaExec) {
-    main = 'ch.softappeal.yass.tutorial.client.web.GenerateTypeScriptContract'
+    main = 'ch.softappeal.yass.tutorial.initiator.web.GenerateTypeScriptContract'
     classpath = sourceSets.tutorial.runtimeClasspath
     standardOutput = new OutputStream() {
         void write(int b) {
@@ -59,24 +60,28 @@ def isWindows() {
 task compileTestTypeScript(type: Exec, dependsOn: 'generateTypeScriptContract') {
     workingDir 'src/test/ts'
     if (isWindows()) {
-        commandLine 'cmd', '/c', 'tsc', '--out', 'all.js', '--noImplicitAny', '--noEmitOnError', '--target', 'ES5', '--sourcemap', 'test'
+        commandLine 'cmd', '/c', 'tsc'
     } else {
         environment PATH: '/usr/local/bin'
-        commandLine '/usr/local/bin/tsc', '--out', 'all.js', '--noImplicitAny', '--noEmitOnError', '--target', 'ES5', '--sourcemap', 'test'
+        commandLine '/usr/local/bin/tsc'
     }
 }
 
 task compileTutorialTypeScript(type: Exec, dependsOn: 'generateTypeScriptContract') {
     workingDir 'src/tutorial/ts'
     if (isWindows()) {
-        commandLine 'cmd', '/c', 'tsc', '--out', 'all.js', '--noImplicitAny', '--noEmitOnError', '--target', 'ES5', '--sourcemap', 'tutorial'
+        commandLine 'cmd', '/c', 'tsc'
     } else {
         environment PATH: '/usr/local/bin'
-        commandLine '/usr/local/bin/tsc', '--out', 'all.js', '--noImplicitAny', '--noEmitOnError', '--target', 'ES5', '--sourcemap', 'tutorial'
+        commandLine '/usr/local/bin/tsc'
     }
 }
 
 task compileAllTypeScript(dependsOn: ['compileTestTypeScript', 'compileTutorialTypeScript'])
+
+tasks.withType(Javadoc) {
+    options.addStringOption('Xdoclint:all,-missing', '-quiet')
+}
 
 jar {
     from "readme.md"
@@ -127,14 +132,15 @@ task collectJars(type: Copy) {
 }
 
 task dist << {
-    ant.copy(todir: "build/server") {
+    ant.copy(todir: "build/acceptor") {
+        fileset(dir: "docker")
         fileset(dir: ".", includes: "src/*/ts/")
     }
-    ant.copy(todir: "build/server/libs") {
+    ant.copy(todir: "build/acceptor/libs") {
         fileset(dir: "build/allJars", includes: "*.jar")
     }
-    ant.jar(destfile: "build/server/libs/yass.jar", basedir: "build/classes/main")
-    ant.jar(destfile: "build/server/libs/tutorial.jar", basedir: "build/classes/tutorial")
+    ant.jar(destfile: "build/acceptor/libs/yass.jar", basedir: "build/classes/main")
+    ant.jar(destfile: "build/acceptor/libs/tutorial.jar", basedir: "build/classes/tutorial")
     if (project.hasProperty('signing.password')) {
         def signPrefix = "build/libs/${project.name}-${version}"
         signing.sign(file("${signPrefix}.pom"))
@@ -179,8 +185,8 @@ task todo << {
 
 idea {
     project {
-        jdkName = '1.7'
-        languageLevel = '1.7'
+        jdkName = '1.8'
+        languageLevel = '1.8'
     }
     module {
         testSourceDirs += sourceSets.tutorial.java.srcDirs
