@@ -1,19 +1,15 @@
 package ch.softappeal.yass.core.remote.session;
 
 import ch.softappeal.yass.util.Check;
+import ch.softappeal.yass.util.Nullable;
 
 import java.util.concurrent.Executor;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicReference;
 
-public final class Reconnector {
+public final class Reconnector<S extends Session> {
 
-    private Reconnector() {
-        // disable
-    }
-
-    public interface Connector {
+    @FunctionalInterface public interface Connector {
         /**
          * @throws Exception note: will be ignored
          */
@@ -23,56 +19,71 @@ public final class Reconnector {
     /**
      * @param executor must interrupt it's threads to terminate reconnects (use {@link ExecutorService#shutdownNow()})
      */
-    public static void start(
+    public Reconnector(
         final Executor executor, final long initialDelaySeconds, final long delaySeconds,
         final SessionFactory sessionFactory, final Connector connector
     ) {
         Check.notNull(sessionFactory);
         Check.notNull(connector);
-        final AtomicReference<Session> session = new AtomicReference<>(null);
-        final SessionFactory factory = new SessionFactory() {
-            @Override public Session create(final SessionClient sessionClient) throws Exception {
-                final Session s = sessionFactory.create(sessionClient);
-                session.set(s);
-                return s;
-            }
+        final SessionFactory proxySessionFactory = connection -> {
+            final Session session = sessionFactory.create(connection);
+            this.session = session;
+            return session;
         };
-        executor.execute(new Runnable() {
-            @Override public void run() {
-                if (initialDelaySeconds > 0) {
+        executor.execute(() -> {
+            if (initialDelaySeconds > 0) {
+                try {
+                    TimeUnit.SECONDS.sleep(initialDelaySeconds);
+                } catch (final InterruptedException ignore) {
+                    return;
+                }
+            }
+            while (!Thread.interrupted()) {
+                final Session session = this.session;
+                if ((session == null) || session.isClosed()) {
+                    this.session = null;
                     try {
-                        TimeUnit.SECONDS.sleep(initialDelaySeconds);
-                    } catch (final InterruptedException ignore) {
-                        return;
+                        connector.connect(proxySessionFactory);
+                    } catch (final Exception ignore) {
+                        // empty
                     }
                 }
-                while (!Thread.interrupted()) {
-                    final Session s = session.get();
-                    if ((s == null) || s.isClosed()) {
-                        try {
-                            connector.connect(factory);
-                        } catch (final Exception ignore) {
-                            // empty
-                        }
-                    }
-                    try {
-                        TimeUnit.SECONDS.sleep(delaySeconds);
-                    } catch (final InterruptedException ignore) {
-                        return;
-                    }
+                try {
+                    TimeUnit.SECONDS.sleep(delaySeconds);
+                } catch (final InterruptedException ignore) {
+                    return;
                 }
             }
         });
     }
 
     /**
-     * @see #start(Executor, long, long, SessionFactory, Connector)
+     * @see #Reconnector(Executor, long, long, SessionFactory, Connector)
      */
-    public static void start(
+    public Reconnector(
         final Executor executor, final long delaySeconds,
         final SessionFactory sessionFactory, final Connector connector
     ) {
-        start(executor, 0, delaySeconds, sessionFactory, connector);
+        this(executor, 0, delaySeconds, sessionFactory, connector);
+    }
+
+    private volatile @Nullable Session session = null;
+
+    /**
+     * @return current {@link Session}
+     * @throws SessionClosedException if no active session
+     */
+    @SuppressWarnings("unchecked")
+    public S session() throws SessionClosedException {
+        final Session session = this.session;
+        if (session == null) {
+            throw new SessionClosedException();
+        }
+        return (S)session;
+    }
+
+    public boolean connected() {
+        return (session != null);
     }
 
 }
