@@ -3,6 +3,7 @@ package ch.softappeal.yass.ts;
 import ch.softappeal.yass.Version;
 import ch.softappeal.yass.core.remote.ContractId;
 import ch.softappeal.yass.core.remote.MethodMapper;
+import ch.softappeal.yass.core.remote.Services;
 import ch.softappeal.yass.core.remote.SimpleMethodMapper;
 import ch.softappeal.yass.serialize.fast.AbstractJsFastSerializer;
 import ch.softappeal.yass.serialize.fast.ClassTypeHandler;
@@ -16,12 +17,12 @@ import ch.softappeal.yass.util.Reflect;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
+import java.lang.reflect.Parameter;
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedHashMap;
@@ -29,16 +30,20 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.SortedMap;
+import java.util.stream.Collectors;
 
+/**
+ * You must use the "-parameters" option for javac to get the real method parameter names.
+ */
 public final class ContractGenerator extends Generator {
 
     private final String rootPackage;
     private final LinkedHashMap<Class<?>, Integer> type2id = new LinkedHashMap<>();
     private final SortedMap<Integer, TypeHandler> id2typeHandler;
     private final Set<Class<?>> visitedClasses = new HashSet<>();
-    private final @Nullable MethodMapper.Factory methodMapperFactory;
     private final Map<Class<?>, String> java2tsBaseType = new HashMap<>();
     private final String contractModuleName;
+    private MethodMapper.Factory methodMapperFactory;
 
     private void checkType(final Class<?> type) {
         if (!type.getCanonicalName().startsWith(rootPackage)) {
@@ -55,7 +60,7 @@ public final class ContractGenerator extends Generator {
         return type.getCanonicalName().substring(rootPackage.length());
     }
 
-    private interface TypeGenerator {
+    @FunctionalInterface private interface TypeGenerator {
         void generateType(String name);
     }
 
@@ -76,21 +81,19 @@ public final class ContractGenerator extends Generator {
     }
 
     private void generateEnum(final Class<? extends Enum<?>> type) {
-        generateType(type, new TypeGenerator() {
-            @Override public void generateType(final String name) {
-                tabsln("export class %s extends yass.Enum {", name);
-                inc();
-                for (final Enum<?> e : type.getEnumConstants()) {
-                    tabsln("static %s = new %s(%s, '%s');", e.name(), name, e.ordinal(), e.name());
-                }
-                tabsln("static TYPE_DESC = yass.enumDesc(%s, %s);", type2id.get(type), name);
-                dec();
-                tabsln("}");
+        generateType(type, name -> {
+            tabsln("export class %s extends yass.Enum {", name);
+            inc();
+            for (final Enum<?> e : type.getEnumConstants()) {
+                tabsln("static %s = new %s(%s, '%s');", e.name(), name, e.ordinal(), e.name());
             }
+            tabsln("static TYPE_DESC = yass.enumDesc(%s, %s);", type2id.get(type), name);
+            dec();
+            tabsln("}");
         });
     }
 
-    private static final Set<Type> ROOT_CLASSES = new HashSet<>(Arrays.<Type>asList(
+    private static final Set<Type> ROOT_CLASSES = new HashSet<>(Arrays.asList(
         Object.class,
         Exception.class,
         RuntimeException.class,
@@ -151,31 +154,29 @@ public final class ContractGenerator extends Generator {
             generateClass(sc);
         }
         final Class<?> superClass = sc;
-        generateType(type, new TypeGenerator() {
-            @Override public void generateType(final String name) {
-                tabsln(
-                    "export %sclass %s extends %s {",
-                    (Modifier.isAbstract(type.getModifiers()) ? "abstract " : ""), name, (superClass == null) ? "yass.Type" : (contractModuleName + jsType(superClass))
-                );
-                inc();
-                for (final Field field : Reflect.ownFields(type)) {
-                    tabsln("%s: %s;", field.getName(), type(field.getGenericType()));
-                }
-                final Integer id = type2id.get(type);
-                if (id != null) {
-                    tabs("static TYPE_DESC = yass.classDesc(%s, %s", id, name);
-                    inc();
-                    for (final ClassTypeHandler.FieldDesc fieldDesc : ((ClassTypeHandler)id2typeHandler.get(id)).fieldDescs()) {
-                        println(",");
-                        tabs("new yass.FieldDesc(%s, '%s', %s)", fieldDesc.id, fieldDesc.handler.field.getName(), typeDescOwner(fieldDesc));
-                    }
-                    println();
-                    dec();
-                    tabsln(");");
-                }
-                dec();
-                tabsln("}");
+        generateType(type, name -> {
+            tabsln(
+                "export %sclass %s extends %s {",
+                (Modifier.isAbstract(type.getModifiers()) ? "abstract " : ""), name, (superClass == null) ? "yass.Type" : (contractModuleName + jsType(superClass))
+            );
+            inc();
+            for (final Field field : Reflect.ownFields(type)) {
+                tabsln("%s: %s;", field.getName(), type(field.getGenericType()));
             }
+            final Integer id = type2id.get(type);
+            if (id != null) {
+                tabs("static TYPE_DESC = yass.classDesc(%s, %s", id, name);
+                inc();
+                for (final ClassTypeHandler.FieldDesc fieldDesc : ((ClassTypeHandler)id2typeHandler.get(id)).fieldDescs()) {
+                    println(",");
+                    tabs("new yass.FieldDesc(%s, '%s', %s)", fieldDesc.id, fieldDesc.handler.field.getName(), typeDescOwner(fieldDesc));
+                }
+                println();
+                dec();
+                tabsln(");");
+            }
+            dec();
+            tabsln("}");
         });
     }
 
@@ -188,62 +189,51 @@ public final class ContractGenerator extends Generator {
         }
     }
 
-    private static List<ServiceDesc> getServiceDescs(final Class<?> services) throws Exception {
+    private static List<ServiceDesc> getServiceDescs(final Services services) throws Exception {
         final List<ServiceDesc> serviceDescs = new ArrayList<>();
-        for (final Field field : services.getFields()) {
-            if (Modifier.isStatic(field.getModifiers()) && (field.getType() == ContractId.class)) {
-                serviceDescs.add(new ServiceDesc(field.getName(), (ContractId<?>)field.get(null)));
+        for (final Field field : services.getClass().getFields()) {
+            if (!Modifier.isStatic(field.getModifiers()) && (field.getType() == ContractId.class)) {
+                serviceDescs.add(new ServiceDesc(field.getName(), (ContractId<?>)field.get(services)));
             }
         }
-        Collections.sort(serviceDescs, new Comparator<ServiceDesc>() {
-            @Override public int compare(final ServiceDesc serviceDesc1, final ServiceDesc serviceDesc2) {
-                return ((Integer)serviceDesc1.contractId.id).compareTo((Integer)serviceDesc2.contractId.id);
-            }
-        });
+        Collections.sort(
+            serviceDescs,
+            (serviceDesc1, serviceDesc2) -> ((Integer)serviceDesc1.contractId.id).compareTo((Integer)serviceDesc2.contractId.id)
+        );
         return serviceDescs;
     }
 
-    private static Set<Class<?>> getInterfaces(final @Nullable Class<?> services) throws Exception {
+    private static Set<Class<?>> getInterfaces(final @Nullable Services services) throws Exception {
         if (services == null) {
             return new HashSet<>();
         }
-        final Set<Class<?>> interfaces = new HashSet<>();
-        for (final ServiceDesc serviceDesc : getServiceDescs(services)) {
-            interfaces.add(serviceDesc.contractId.contract);
-        }
-        return interfaces;
+        return getServiceDescs(services).stream().map(serviceDesc -> serviceDesc.contractId.contract).collect(Collectors.toSet());
     }
 
     private static Method[] getMethods(final Class<?> type) {
         final Method[] methods = type.getMethods();
-        Arrays.sort(methods, new Comparator<Method>() {
-            @Override public int compare(final Method method1, final Method method2) {
-                return method1.getName().compareTo(method2.getName());
-            }
-        });
+        Arrays.sort(methods, (method1, method2) -> method1.getName().compareTo(method2.getName()));
         return methods;
     }
 
     private void generateInterface(final Class<?> type) {
         checkType(type);
         final Method[] methods = getMethods(type);
-        if (methodMapperFactory == null) {
-            throw new IllegalArgumentException("methodMapperFactory must be specified if there are services");
-        }
-        SimpleMethodMapper.FACTORY.create(type); // checks for overloaded methods
-        final MethodMapper methodMapper = methodMapperFactory.create(type);
+        SimpleMethodMapper.FACTORY.create(type); // checks for overloaded methods (JavaScript restriction)
+        final MethodMapper methodMapper = SimpleMethodMapper.FACTORY.create(type);
         generateType(type, new TypeGenerator() {
             void generateInterface(final String name, final boolean proxy) {
                 tabsln("export interface %s {", name + (proxy ? "_PROXY" : ""));
                 inc();
                 for (final Method method : methods) {
                     tabs("%s(", method.getName());
-                    int param = 0;
-                    for (final Type paramType : method.getGenericParameterTypes()) {
-                        if (param != 0) {
+                    boolean first = true;
+                    for (final Parameter parameter : method.getParameters()) {
+                        if (!first) {
                             print(", ");
                         }
-                        print("param%s: %s", param++, type(paramType));
+                        first = false;
+                        print("%s: %s", parameter.getName(), type(parameter.getParameterizedType()));
                     }
                     print("): ");
                     if (methodMapper.mapMethod(method).oneWay) {
@@ -283,11 +273,11 @@ public final class ContractGenerator extends Generator {
         });
     }
 
-    private void generateServices(final @Nullable Class<?> services) throws Exception {
+    private void generateServices(final @Nullable Services services, final String role) throws Exception {
         if (services == null) {
             return;
         }
-        tabsln("export namespace %s {", jsType(services));
+        tabsln("export namespace %s {", role);
         inc();
         for (final ServiceDesc serviceDesc : getServiceDescs(services)) {
             final String name = contractModuleName + jsType(serviceDesc.contractId.contract);
@@ -301,27 +291,15 @@ public final class ContractGenerator extends Generator {
         println();
     }
 
-    public static final String CLIENT_SERVICES = "ClientServices";
-    public static final String SERVER_SERVICES = "ServerServices";
-
-    private @Nullable Class<?> getServicesClass(final String servicesClass) {
-        try {
-            return Class.forName(this.rootPackage + servicesClass);
-        } catch (final ClassNotFoundException ignore) {
-            return null;
-        }
-    }
-
     /**
-     * @param rootPackage Must contain the optional classes {@link #CLIENT_SERVICES} and {@link #SERVER_SERVICES} with static fields of type {@link ContractId}.
-     * @param methodMapperFactory Optional if there are no services. You must provide a factory that doesn't allow overloading due to JavaScript restrictions.
      * @param includePath path to base types or yass module
      */
     @SuppressWarnings("unchecked")
     public ContractGenerator(
         final Package rootPackage,
         final AbstractJsFastSerializer serializer,
-        final @Nullable MethodMapper.Factory methodMapperFactory,
+        final @Nullable Services initiator,
+        final @Nullable Services acceptor,
         final String includePath,
         final String contractModuleName,
         final @Nullable Map<Class<?>, String> java2tsBaseType,
@@ -329,19 +307,15 @@ public final class ContractGenerator extends Generator {
     ) throws Exception {
         super(contractFilePath);
         this.rootPackage = rootPackage.getName() + '.';
-        this.methodMapperFactory = methodMapperFactory;
         if (java2tsBaseType != null) {
-            for (final Map.Entry<Class<?>, String> entry : java2tsBaseType.entrySet()) {
-                this.java2tsBaseType.put(Check.notNull(entry.getKey()), Check.notNull(entry.getValue()));
-            }
+            java2tsBaseType.forEach((java, ts) -> this.java2tsBaseType.put(Check.notNull(java), Check.notNull(ts)));
         }
         id2typeHandler = serializer.id2typeHandler();
-        for (final Map.Entry<Integer, TypeHandler> entry : id2typeHandler.entrySet()) {
-            final int id = entry.getKey();
+        id2typeHandler.forEach((id, typeHandler) -> {
             if (id >= AbstractJsFastSerializer.FIRST_ID) {
-                type2id.put(entry.getValue().type, id);
+                type2id.put(typeHandler.type, id);
             }
-        }
+        });
         tabsln("/// <reference path='%s'/>", Check.notNull(includePath));
         println();
         this.contractModuleName = Check.notNull(contractModuleName) + '.';
@@ -350,32 +324,24 @@ public final class ContractGenerator extends Generator {
         inc();
         tabsln("export const GENERATED_BY_YASS_VERSION = '%s';", Version.VALUE);
         println();
-        for (final TypeHandler typeHandler : id2typeHandler.values()) {
-            final Class<?> type = typeHandler.type;
-            if (type.isEnum()) {
-                generateEnum((Class<Enum<?>>)type);
-            }
+        id2typeHandler.values().stream().map(typeHandler -> typeHandler.type).filter(Class::isEnum).forEach(type -> generateEnum((Class<Enum<?>>)type));
+        id2typeHandler.values().stream().filter(typeHandler -> typeHandler instanceof ClassTypeHandler).forEach(typeHandler -> generateClass(typeHandler.type));
+        if ((initiator != null) && (acceptor != null) && (initiator.methodMapperFactory != acceptor.methodMapperFactory)) {
+            throw new IllegalArgumentException("initiator and acceptor must have same methodMapperFactory");
         }
-        for (final TypeHandler typeHandler : id2typeHandler.values()) {
-            if (typeHandler instanceof ClassTypeHandler) {
-                generateClass(typeHandler.type);
-            }
+        if (initiator != null) {
+            methodMapperFactory = initiator.methodMapperFactory;
         }
-        @Nullable final Class<?> clientServices = getServicesClass(CLIENT_SERVICES);
-        @Nullable final Class<?> serverServices = getServicesClass(SERVER_SERVICES);
-        final Set<Class<?>> interfaceSet = getInterfaces(clientServices);
-        interfaceSet.addAll(getInterfaces(serverServices));
+        if (acceptor != null) {
+            methodMapperFactory = acceptor.methodMapperFactory;
+        }
+        final Set<Class<?>> interfaceSet = getInterfaces(initiator);
+        interfaceSet.addAll(getInterfaces(acceptor));
         final List<Class<?>> interfaceList = new ArrayList<>(interfaceSet);
-        Collections.sort(interfaceList, new Comparator<Class<?>>() {
-            @Override public int compare(final Class<?> type1, final Class<?> type2) {
-                return type1.getCanonicalName().compareTo(type2.getCanonicalName());
-            }
-        });
-        for (final Class<?> aInterface : interfaceList) {
-            generateInterface(aInterface);
-        }
-        generateServices(clientServices);
-        generateServices(serverServices);
+        Collections.sort(interfaceList, (type1, type2) -> type1.getCanonicalName().compareTo(type2.getCanonicalName()));
+        interfaceList.forEach(this::generateInterface);
+        generateServices(initiator, "initiator");
+        generateServices(acceptor, "acceptor");
         tabs("export const SERIALIZER = new yass.JsFastSerializer(");
         inc();
         boolean first = true;
