@@ -1,7 +1,6 @@
 package ch.softappeal.yass.core.test;
 
 import ch.softappeal.yass.core.Interceptor;
-import ch.softappeal.yass.core.Interceptors;
 import ch.softappeal.yass.core.Invocation;
 import ch.softappeal.yass.core.remote.OneWay;
 import ch.softappeal.yass.util.Check;
@@ -31,9 +30,7 @@ public class InvokeTest {
     public interface TestService {
         @Tag(0) void nothing();
         @Tag(11) int divide(int a, int b) throws DivisionByZeroException;
-        @Tag(22) void throwError();
         @Tag(33) @OneWay void oneWay(int sleepMillis);
-        @Tag(34) void delay(int milliSeconds);
     }
 
     public static void println(final String name, final String type, final Object message) {
@@ -53,9 +50,6 @@ public class InvokeTest {
             }
             return (a / b);
         }
-        @Override public void throwError() {
-            throw new Error("throwError");
-        }
         @Override public void oneWay(final int sleepMillis) {
             if (sleepMillis < 0) {
                 return;
@@ -67,13 +61,6 @@ public class InvokeTest {
                 throw new RuntimeException(e);
             }
         }
-        @Override public void delay(final int milliSeconds) {
-            try {
-                TimeUnit.MILLISECONDS.sleep(milliSeconds);
-            } catch (final InterruptedException e) {
-                throw new RuntimeException(e);
-            }
-        }
     }
 
     private static final class Logger implements Interceptor {
@@ -81,15 +68,15 @@ public class InvokeTest {
         Logger(final String name) {
             this.name = Check.notNull(name);
         }
-        @Override public Object invoke(final Method method, final @Nullable Object[] arguments, final Invocation invocation) throws Throwable {
+        @Override public Object invoke(final Method method, final @Nullable Object[] arguments, final Invocation invocation) throws Exception {
             println(name, "entry", method.getName() + ' ' + Arrays.deepToString(arguments));
             try {
                 final Object result = invocation.proceed();
                 println(name, "exit", result);
                 return result;
-            } catch (final Throwable t) {
-                println(name, "exception", t);
-                throw t;
+            } catch (final Exception e) {
+                println(name, "exception", e);
+                throw e;
             }
         }
     }
@@ -98,13 +85,11 @@ public class InvokeTest {
     private static final AtomicReference<Method> METHOD = new AtomicReference<>();
     private static final AtomicReference<Object[]> ARGUMENTS = new AtomicReference<>();
 
-    public static final Interceptor PRINTLN_AFTER = new Interceptor() {
-        @Override public Object invoke(final Method method, @Nullable final Object[] arguments, final Invocation invocation) throws Throwable {
-            try {
-                return invocation.proceed();
-            } finally {
-                System.out.println();
-            }
+    public static final Interceptor PRINTLN_AFTER = (method, arguments, invocation) -> {
+        try {
+            return invocation.proceed();
+        } finally {
+            System.out.println();
         }
     };
 
@@ -117,24 +102,20 @@ public class InvokeTest {
         }
     }
 
-    public static final Interceptor CLIENT_INTERCEPTOR = Interceptors.composite(
-        new Interceptor() {
-            @Override public Object invoke(final Method method, @Nullable final Object[] arguments, final Invocation invocation) throws Throwable {
-                Assert.assertTrue(COUNTER.incrementAndGet() == 1);
-                checkArguments(method, arguments);
-                return invocation.proceed();
-            }
+    public static final Interceptor CLIENT_INTERCEPTOR = Interceptor.composite(
+        (method, arguments, invocation) -> {
+            Assert.assertTrue(COUNTER.incrementAndGet() == 1);
+            checkArguments(method, arguments);
+            return invocation.proceed();
         },
         new Logger("client")
     );
 
-    public static final Interceptor SERVER_INTERCEPTOR = Interceptors.composite(
-        new Interceptor() {
-            @Override public Object invoke(final Method method, @Nullable final Object[] arguments, final Invocation invocation) throws Throwable {
-                Assert.assertTrue(COUNTER.incrementAndGet() == 2);
-                checkArguments(method, arguments);
-                return invocation.proceed();
-            }
+    public static final Interceptor SERVER_INTERCEPTOR = Interceptor.composite(
+        (method, arguments, invocation) -> {
+            Assert.assertTrue(COUNTER.incrementAndGet() == 2);
+            checkArguments(method, arguments);
+            return invocation.proceed();
         },
         new Logger("server")
     );
@@ -159,45 +140,23 @@ public class InvokeTest {
      * @param testService must use {@link #CLIENT_INTERCEPTOR} and {@link #SERVER_INTERCEPTOR}
      */
     public static void invoke(final TestService testService) throws InterruptedException {
-        invoke("nothing", new Class<?>[] {}, null, new Runnable() {
-            @Override public void run() {
-                testService.nothing();
+        invoke("nothing", new Class<?>[] {}, null, testService::nothing);
+        invoke("divide", new Class<?>[] {int.class, int.class}, new Object[] {12, 3}, () -> {
+            try {
+                Assert.assertTrue(testService.divide(12, 3) == 4);
+            } catch (final DivisionByZeroException e) {
+                Assert.fail();
             }
         });
-        invoke("divide", new Class<?>[] {int.class, int.class}, new Object[] {12, 3}, new Runnable() {
-            @Override public void run() {
-                try {
-                    Assert.assertTrue(testService.divide(12, 3) == 4);
-                } catch (final DivisionByZeroException e) {
-                    Assert.fail();
-                }
+        invoke("divide", new Class<?>[] {int.class, int.class}, new Object[] {123, 0}, () -> {
+            try {
+                testService.divide(123, 0);
+                Assert.fail();
+            } catch (final DivisionByZeroException e) {
+                Assert.assertTrue(e.value == 123);
             }
         });
-        invoke("divide", new Class<?>[] {int.class, int.class}, new Object[] {123, 0}, new Runnable() {
-            @Override public void run() {
-                try {
-                    testService.divide(123, 0);
-                    Assert.fail();
-                } catch (final DivisionByZeroException e) {
-                    Assert.assertTrue(e.value == 123);
-                }
-            }
-        });
-        invoke("throwError", new Class<?>[] {}, null, new Runnable() {
-            @Override public void run() {
-                try {
-                    testService.throwError();
-                    Assert.fail();
-                } catch (final Error e) {
-                    Assert.assertEquals("throwError", e.getMessage());
-                }
-            }
-        });
-        invoke("oneWay", new Class<?>[] {int.class}, new Object[] {100}, new Runnable() {
-            @Override public void run() {
-                testService.oneWay(100);
-            }
-        });
+        invoke("oneWay", new Class<?>[] {int.class}, new Object[] {100}, () -> testService.oneWay(100));
         TimeUnit.MILLISECONDS.sleep(200L);
     }
 
