@@ -11,8 +11,8 @@ public interface Invocation {
     /**
      * Proceeds with the invocation.
      * @return the result of the invocation
-     * @throws Throwable the exception of the invocation
+     * @throws Exception the exception of the invocation
      */
-    @Nullable Object proceed() throws Throwable;
+    @Nullable Object proceed() throws Exception;
 
 }
