@@ -3,7 +3,7 @@ package ch.softappeal.yass.core.remote.session;
 public interface Connection {
 
     /**
-     * Called if a packet has to be written out. Must be thread safe.
+     * Called if a packet has to be written out.
      */
     void write(Packet packet) throws Exception;
 
