@@ -1,27 +1,20 @@
 package ch.softappeal.yass.core.remote.session.test;
 
 import ch.softappeal.yass.core.Interceptor;
-import ch.softappeal.yass.core.Interceptors;
-import ch.softappeal.yass.core.Invocation;
 import ch.softappeal.yass.core.remote.Server;
-import ch.softappeal.yass.core.remote.Service;
-import ch.softappeal.yass.core.remote.TaggedMethodMapper;
 import ch.softappeal.yass.core.remote.session.LocalConnection;
 import ch.softappeal.yass.core.remote.session.Session;
-import ch.softappeal.yass.core.remote.session.SessionClient;
+import ch.softappeal.yass.core.remote.session.SessionClosedException;
 import ch.softappeal.yass.core.remote.session.SessionFactory;
+import ch.softappeal.yass.core.remote.session.SimpleSession;
 import ch.softappeal.yass.core.remote.test.ContractIdTest;
 import ch.softappeal.yass.core.test.InvokeTest;
-import ch.softappeal.yass.transport.TransportSetup;
 import ch.softappeal.yass.transport.socket.SocketHelper;
-import ch.softappeal.yass.transport.test.PacketSerializerTest;
 import ch.softappeal.yass.util.Exceptions;
 import ch.softappeal.yass.util.NamedThreadFactory;
-import ch.softappeal.yass.util.Nullable;
 import org.junit.Assert;
 import org.junit.Test;
 
-import java.lang.reflect.Method;
 import java.util.concurrent.Executor;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
@@ -29,74 +22,66 @@ import java.util.concurrent.TimeUnit;
 
 public class LocalConnectionTest extends InvokeTest {
 
-    private static final Interceptor SESSION_CHECKER = new Interceptor() {
-        @Override public Object invoke(final Method method, @Nullable final Object[] arguments, final Invocation invocation) throws Throwable {
-            Assert.assertNotNull(Session.get());
-            return invocation.proceed();
-        }
-    };
-
-    private static TransportSetup createSetup(
+    private static SessionFactory sessionFactory(
         final boolean invoke,
-        final Executor dispatcherExecutor,
+        final Executor dispatchExecutor,
         final boolean createException,
         final boolean openedException,
         final boolean invokeBeforeOpened
     ) {
-        return new TransportSetup(
-            new Server(
-                TaggedMethodMapper.FACTORY,
-                new Service(ContractIdTest.ID, new TestServiceImpl(), invoke ? SESSION_CHECKER : Interceptors.composite(SESSION_CHECKER, SERVER_INTERCEPTOR))
-            ),
-            dispatcherExecutor,
-            PacketSerializerTest.SERIALIZER,
-            new SessionFactory() {
-                @Override public Session create(final SessionClient sessionClient) throws Exception {
-                    if (createException) {
-                        throw new Exception("create failed");
-                    }
-                    if (invokeBeforeOpened) {
-                        sessionClient.proxy(ContractIdTest.ID).nothing();
-                    }
-                    return new Session(sessionClient) {
-                        @Override public void opened() throws Exception {
-                            println("", "opened", hashCode());
-                            if (openedException) {
-                                throw new Exception("opened failed");
-                            }
-                            if (invoke) {
-                                try (Session session = this) {
-                                    InvokeTest.invoke(session.proxy(
-                                        ContractIdTest.ID,
-                                        invoke ? Interceptors.composite(PRINTLN_AFTER, SESSION_CHECKER, CLIENT_INTERCEPTOR) : SESSION_CHECKER
-                                    ));
-                                }
-                            }
-                        }
-                        @Override public void closed(final @Nullable Throwable throwable) {
-                            if (invoke) {
-                                Assert.assertNull(throwable);
-                            }
-                            println("", "closed", hashCode() + " " + throwable);
-                        }
-                    };
-                }
+        return connection -> {
+            if (createException) {
+                throw new Exception("create failed");
             }
-        );
+            return new SimpleSession(connection, dispatchExecutor) {
+                {
+                    if (invokeBeforeOpened) {
+                        proxy(ContractIdTest.ID).nothing();
+                    }
+                }
+                @Override protected Server server() {
+                    return new Server(
+                        ContractIdTest.ID.service(new TestServiceImpl(), invoke ? Interceptor.DIRECT : SERVER_INTERCEPTOR)
+                    );
+                }
+                @Override protected void opened() throws Exception {
+                    println("", "opened", hashCode());
+                    if (openedException) {
+                        throw new Exception("opened failed");
+                    }
+                    if (invoke) {
+                        try (Session session = this) {
+                            InvokeTest.invoke(session.proxy(
+                                ContractIdTest.ID,
+                                Interceptor.composite(PRINTLN_AFTER, CLIENT_INTERCEPTOR)
+                            ));
+                        }
+                        try {
+                            proxy(ContractIdTest.ID).nothing();
+                            Assert.fail();
+                        } catch (final SessionClosedException ignored) {
+                            // empty
+                        }
+                    }
+                }
+                @Override protected void closed(final boolean exceptional) {
+                    println("", "closed", hashCode() + " " + exceptional);
+                }
+            };
+        };
     }
 
-    public static TransportSetup createSetup(final boolean invoke, final Executor dispatcherExecutor, final boolean createException) {
-        return createSetup(invoke, dispatcherExecutor, createException, false, false);
+    public static SessionFactory sessionFactory(final boolean invoke, final Executor dispatchExecutor, final boolean createException) {
+        return sessionFactory(invoke, dispatchExecutor, createException, false, false);
     }
 
     @Test public void plain() throws InterruptedException {
-        Assert.assertNull(Session.get());
         final ExecutorService executor = Executors.newCachedThreadPool(new NamedThreadFactory("executor", Exceptions.TERMINATE));
         try {
-            LocalConnection.connect(createSetup(true, executor, false), createSetup(false, executor, false));
+            LocalConnection.connect(sessionFactory(true, executor, false), sessionFactory(false, executor, false));
             TimeUnit.MILLISECONDS.sleep(400L);
             System.out.println();
-            LocalConnection.connect(createSetup(false, executor, false), createSetup(true, executor, false));
+            LocalConnection.connect(sessionFactory(false, executor, false), sessionFactory(true, executor, false));
             TimeUnit.MILLISECONDS.sleep(400L);
         } finally {
             executor.shutdownNow();
@@ -107,7 +92,7 @@ public class LocalConnectionTest extends InvokeTest {
         final ExecutorService executor = Executors.newCachedThreadPool(new NamedThreadFactory("executor", Exceptions.TERMINATE));
         try {
             try {
-                LocalConnection.connect(createSetup(false, executor, false), createSetup(false, executor, true));
+                LocalConnection.connect(sessionFactory(false, executor, false), sessionFactory(false, executor, true));
                 Assert.fail();
             } catch (final RuntimeException e) {
                 Assert.assertEquals(e.getMessage(), "java.lang.Exception: create failed");
@@ -118,9 +103,9 @@ public class LocalConnectionTest extends InvokeTest {
     }
 
     @Test public void openedException() throws InterruptedException {
-        final ExecutorService executor = Executors.newCachedThreadPool(new NamedThreadFactory("executor", Exceptions.TERMINATE));
+        final ExecutorService executor = Executors.newCachedThreadPool(new NamedThreadFactory("executor", Exceptions.STD_ERR));
         try {
-            LocalConnection.connect(createSetup(false, executor, false, false, false), createSetup(false, executor, false, true, false));
+            LocalConnection.connect(sessionFactory(false, executor, false, false, false), sessionFactory(false, executor, false, true, false));
             TimeUnit.MILLISECONDS.sleep(100L);
         } finally {
             SocketHelper.shutdown(executor);
@@ -131,9 +116,9 @@ public class LocalConnectionTest extends InvokeTest {
         final ExecutorService executor = Executors.newCachedThreadPool(new NamedThreadFactory("executor", Exceptions.TERMINATE));
         try {
             try {
-                LocalConnection.connect(createSetup(true, executor, false, false, true), createSetup(false, executor, false));
+                LocalConnection.connect(sessionFactory(true, executor, false, false, true), sessionFactory(false, executor, false));
                 Assert.fail();
-            } catch (final NullPointerException e) {
+            } catch (final SessionClosedException e) {
                 e.printStackTrace();
             }
         } finally {
