@@ -1,15 +1,10 @@
 package ch.softappeal.yass.transport.socket.test;
 
 import ch.softappeal.yass.core.remote.ContractId;
-import ch.softappeal.yass.core.remote.MethodMapper;
 import ch.softappeal.yass.core.remote.OneWay;
 import ch.softappeal.yass.core.remote.Server;
-import ch.softappeal.yass.core.remote.Service;
 import ch.softappeal.yass.core.remote.SimpleMethodMapper;
-import ch.softappeal.yass.core.remote.session.Dispatcher;
-import ch.softappeal.yass.core.remote.session.Session;
-import ch.softappeal.yass.core.remote.session.SessionClient;
-import ch.softappeal.yass.core.remote.session.SessionFactory;
+import ch.softappeal.yass.core.remote.session.SimpleSession;
 import ch.softappeal.yass.serialize.JavaSerializer;
 import ch.softappeal.yass.serialize.Serializer;
 import ch.softappeal.yass.transport.TransportSetup;
@@ -18,7 +13,6 @@ import ch.softappeal.yass.transport.socket.SocketTransport;
 import ch.softappeal.yass.transport.socket.SyncSocketConnection;
 import ch.softappeal.yass.util.Exceptions;
 import ch.softappeal.yass.util.NamedThreadFactory;
-import ch.softappeal.yass.util.Nullable;
 
 import java.net.InetSocketAddress;
 import java.net.SocketAddress;
@@ -32,48 +26,34 @@ public final class AsyncSocketConnectionTest {
         @OneWay void busy();
     }
 
-    private static final ContractId<Busy> BUSY_ID = ContractId.create(Busy.class, 0);
+    private static final ContractId<Busy> BUSY_ID = ContractId.create(Busy.class, 0, SimpleMethodMapper.FACTORY);
     private static final SocketAddress ADDRESS = new InetSocketAddress("localhost", 28947);
     private static final Serializer PACKET_SERIALIZER = JavaSerializer.INSTANCE;
-    private static final MethodMapper.Factory METHOD_MAPPER_FACTORY = SimpleMethodMapper.FACTORY;
 
     public static void main(final String... args) throws InterruptedException {
         final ExecutorService executor = Executors.newCachedThreadPool(new NamedThreadFactory("Executor", Exceptions.STD_ERR));
 
         new SocketTransport(executor, SyncSocketConnection.FACTORY).start(
-            new TransportSetup(
-                new Server(
-                    METHOD_MAPPER_FACTORY,
-                    new Service(BUSY_ID, new Busy() {
-                        @Override public void busy() {
-                            System.out.println("busy");
-                            try {
-                                TimeUnit.MILLISECONDS.sleep(1_000);
-                            } catch (InterruptedException e) {
-                                throw new RuntimeException(e);
-                            }
-                        }
-                    })
-                ),
-                new Dispatcher() {
-                    @Override public void opened(final Session session, final Runnable runnable) {
-                        runnable.run();
-                    }
-                    @Override public void invoke(final Session session, final Server.Invocation invocation, final Runnable runnable) {
-                        runnable.run();
-                    }
-                },
+            TransportSetup.ofPacketSerializer(
                 PACKET_SERIALIZER,
-                new SessionFactory() {
-                    @Override public Session create(final SessionClient sessionClient) throws Exception {
-                        return new Session(sessionClient) {
-                            @Override protected void opened() {
-                                System.out.println("server opened");
-                            }
-                            @Override public void closed(final @Nullable Throwable throwable) {
-                                System.out.println("server closed");
-                            }
-                        };
+                connection -> new SimpleSession(connection, executor) {
+                    @Override protected Server server() {
+                        return new Server(
+                            BUSY_ID.service(() -> {
+                                System.out.println("busy");
+                                try {
+                                    TimeUnit.MILLISECONDS.sleep(1_000);
+                                } catch (InterruptedException e) {
+                                    throw new RuntimeException(e);
+                                }
+                            })
+                        );
+                    }
+                    @Override protected void opened() {
+                        System.out.println("acceptor opened");
+                    }
+                    @Override protected void closed(final boolean exceptional) {
+                        System.out.println("acceptor closed");
                     }
                 }
             ),
@@ -81,30 +61,23 @@ public final class AsyncSocketConnectionTest {
             ADDRESS
         );
 
-        new SocketTransport(executor, AsyncSocketConnection.factory(executor, 1_000)).connect(
-            new TransportSetup(
-                new Server(
-                    METHOD_MAPPER_FACTORY
-                ),
-                executor,
+        new SocketTransport(executor, AsyncSocketConnection.factory(executor, 10)).connect(
+            TransportSetup.ofPacketSerializer(
                 PACKET_SERIALIZER,
-                new SessionFactory() {
-                    @Override public Session create(final SessionClient sessionClient) throws Exception {
-                        return new Session(sessionClient) {
-                            @Override protected void opened() {
-                                System.out.println("client opened");
-                                final Busy busy = proxy(BUSY_ID);
-                                for (int i = 0; i < 10_000; i++) {
-                                    busy.busy();
-                                }
-                                System.out.println("client done");
-                            }
-                            @Override public void closed(final @Nullable Throwable throwable) {
-                                System.out.println("client closed");
-                                throwable.printStackTrace(System.out);
-                                System.exit(1);
-                            }
-                        };
+                connection -> new SimpleSession(connection, executor) {
+                    @Override protected Server server() {
+                        return Server.EMPTY;
+                    }
+                    @Override protected void opened() {
+                        System.out.println("initiator opened");
+                        final Busy busy = proxy(BUSY_ID);
+                        for (int i = 0; i < 10_000; i++) {
+                            busy.busy();
+                        }
+                        System.out.println("initiator done");
+                    }
+                    @Override protected void closed(final boolean exceptional) {
+                        System.out.println("initiator closed: " + exceptional);
                     }
                 }
             ),
