@@ -2,7 +2,6 @@ package ch.softappeal.yass.tutorial.contract;
 
 import ch.softappeal.yass.core.Interceptor;
 import ch.softappeal.yass.core.Invocation;
-import ch.softappeal.yass.util.Exceptions;
 import ch.softappeal.yass.util.Nullable;
 
 import java.lang.reflect.Method;
@@ -13,16 +12,13 @@ public final class UnexpectedExceptionHandler implements Interceptor {
         // disable
     }
 
-    @Override public Object invoke(final Method method, final @Nullable Object[] arguments, final Invocation invocation) throws Throwable {
+    @Override public Object invoke(final Method method, final @Nullable Object[] arguments, final Invocation invocation) throws Exception {
         try {
             return invocation.proceed();
         } catch (final ApplicationException e) { // pass through contract exception
             throw e;
         } catch (final Exception e) { // remap unexpected exception to a contract exception
             throw new SystemException(e.getClass().getName() + ": " + e.getMessage());
-        } catch (final Throwable t) { // terminate on Throwable
-            Exceptions.uncaughtException(Exceptions.TERMINATE, t);
-            return null;
         }
     }
 
